[X]5.验厂时的时间显示
2016-4-24
[X]1. 辅料采购单对应的出库入库
2016-4-28
[X]1.验厂时的外发工厂不显示
[X]2.自购辅料工作台的更改库位
[X]3.自购辅料工作台的清空库存

+++ src/com/fuwei/controller/report/ReportController.java
@@ -782,7 +782,7 @@ public class ReportController extends BaseController {
 			sortList.add(sort_factory);
 			
 			
-			HashMap<Factory,HashMap<FuliaoType,Double> > result = fuliaoPurchaseOrderService.fuliao_purchase_report(start_time_d, end_time_d,
+			HashMap<Factory,HashMap<FuliaoType,Integer> > result = fuliaoPurchaseOrderService.fuliao_purchase_report(start_time_d, end_time_d,
 					factoryId, sortList);
 			request.setAttribute("result", result);
 			request.setAttribute("start_time", start_time_d);
